+++ packages/lib/components/select/select.stories.ts
@@ -2,12 +2,12 @@ import { withActions } from '@storybook/addon-actions/decorator';
 import type { Meta, StoryObj } from '@storybook/web-components';
 import { html } from 'lit';
 
-import './dropdown';
+import './select';
 import './option';
 import '../icon';
 
 export default {
-  title: 'Components/Dropdown',
+  title: 'Components/Select',
   parameters: {
     actions: {
       handles: ['change'],
@@ -21,11 +21,11 @@ export const Default: StoryObj = {
 <label class="cx-form-field">
   <div class="cx-form-field__label">Countries</div>
   <div class="cx-form-field__input-container">
-    <cx-dropdown>
+    <cx-select>
       <cx-option value="norway">Norway</cx-option>
       <cx-option value="romania">Romania</cx-option>
       <cx-option value="denmark">Denmark</cx-option>
-    </cx-dropdown>
+    </cx-select>
   </div>
 </label>
   `,
@@ -36,11 +36,11 @@ export const WithValue: StoryObj = {
 <label class="cx-form-field">
   <div class="cx-form-field__label">Countries</div>
   <div class="cx-form-field__input-container">
-    <cx-dropdown value="romania">
+    <cx-select value="romania">
       <cx-option value="norway">Norway</cx-option>
       <cx-option value="romania">Romania</cx-option>
       <cx-option value="denmark">Denmark</cx-option>
-    </cx-dropdown>
+    </cx-select>
   </div>
 </label>
   `,
@@ -51,11 +51,11 @@ export const Required: StoryObj = {
 <label class="cx-form-field">
   <div class="cx-form-field__label">Countries</div>
   <div class="cx-form-field__input-container">
-    <cx-dropdown required aria-describedby="error-text">
+    <cx-select required aria-describedby="error-text">
       <cx-option value="norway">Norway</cx-option>
       <cx-option value="romania">Romania</cx-option>
       <cx-option value="denmark">Denmark</cx-option>
-    </cx-dropdown>
+    </cx-select>
   </div>
 
   <div class="cx-form-field__error" aria-live="polite" id="error-text">