+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.0.9 - xx.01.2025
 
-- Lagt til dropdown-komponent
+- Lagt til select-komponent
 
 # 0.0.8 - 13.01.2025